requirements.txt
torch>=2.0.0
transformers>=4.30.0
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.2.0
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
jupyter>=1.0.0
nbformat>=5.7.0
boto3>=1.26.0
opensearch-py>=2.0.0
requests-aws4auth>=1.1.0
gudhi>=3.7.0
tqdm>=4.65.0
peft>=0.4.0
python-dotenv>=1.0.0
awscli>=1.29.0
